✨ (onboarding): Enable routing for WireBook service (#2)
* ✨ (onboarding): Enable routing for WireBook service

Add routing capability to the WireBook service provider. This change allows the service to handle web routes, enhancing its functionality and integration within the application.

* ✨ (routing): Update story route to use StoryController

Refactor the story route to utilize the StoryController for better organization and maintainability. This change enhances the routing structure and prepares for future feature expansions.

* ✨ (viewport): Update view path for viewport component

Refactor the view path for the viewport component to improve organization and clarity. This change enhances the structure of the codebase and aligns with the updated naming conventions.

* 🚚 (routes): Remove unused route definitions from service provider

Eliminate unnecessary route macros from the WireBookServiceProvider. This cleanup helps streamline the codebase and improves maintainability by removing outdated routing logic.

* ✨ (routing): Add wirebook route registration

Introduce a new route macro for the WireBook package to simplify route registration in the application. This change allows for easy access to the dashboard and story views, enhancing the onboarding experience for users.

* ✨ (config): Enable WireBook feature via environment variable

Add a configuration option to enable or disable the WireBook feature using an environment variable. This allows for better control over the feature's availability in different environments.

* 🚚 (routes): Simplify WireBook route registration

Refactor the route definitions for WireBook by invoking the macro defined in the service provider. This change streamlines the routing process and enhances maintainability by reducing the amount of code needed for route registration.

* 📝 (docs): Update README for improved clarity and formatting

Refine the structure and formatting of the README.md file. Adjust badge links for consistency and enhance readability by breaking long lines into shorter ones. Update vendor publish tags for better clarity.

* ⬆️ (composer): Upgrade PHP requirement to ^8.3
Update the PHP version requirement in composer.json to ensure compatibility with the latest features and improvements in PHP 8.3. This change helps maintain the project's dependency management and prepares for future updates.

* 🔧 (phpstan): Upgrade PHP version to 8.3

Updated the PHP version in the phpstan configuration from 8.2 to 8.3 to ensure compatibility with the latest features and improvements. This change helps maintain the project's alignment with current PHP standards.

* 🔧 (ci): Update test matrix for Laravel and PHP versions

Enhance the CI configuration by updating the PHP versions to include 8.4 and adjusting the Laravel versions to support 12.*, 11.*, and 10.*. This change ensures compatibility with the latest features and improvements in the framework.